requirements.txt
baml-py
python-dotenv
pydantic